requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
json_repair==0.30.0
numpy==1.24.4
openai==1.77.0
pandas==2.0.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tiktoken==0.7.0
timeout-decorator==0.5.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
